requirements.txt
smbus2>=0.4.3
RPi.GPIO>=0.7.0
